+++ src/bespokeasm/assembler/pretty_printer/__init__.py
@@ -3,18 +3,26 @@ from bespokeasm.assembler.model import AssemblerModel
 
 
 class PrettyPrinterBase:
-    def __init__(self, line_objs:  list[LineObject], model: AssemblerModel) -> None:
+    def __init__(
+                self,
+                line_objs: list[LineObject],
+                model: AssemblerModel,
+                instruction_indent: int = 0,
+            ) -> None:
         self._line_objs = line_objs
         self._model = model
         # scan though the line objects to get max widths
         line_num_max = 0
         instruction_max_width = 0
         self._max_byte_count = 0
+        self._max_comment_width = 0
         for lo in line_objs:
             if lo.line_id.line_num > line_num_max:
                 line_num_max = lo.line_id.line_num
-            if len(lo.instruction) > instruction_max_width:
-                instruction_max_width = len(lo.instruction)
+            if len(lo.instruction) + instruction_indent > instruction_max_width:
+                instruction_max_width = len(lo.instruction) + instruction_indent
+            if len(lo.comment) > self._max_comment_width:
+                self._max_comment_width = len(lo.comment)
             if isinstance(lo, LineWithBytes):
                 if len(lo.get_bytes()) > self._max_byte_count:
                     self._max_byte_count = len(lo.get_bytes())
@@ -31,7 +39,7 @@ class PrettyPrinterBase:
 
     @property
     def max_line_num_width(self) -> int:
-        return self._max_line_num_width
+        return max(self._max_line_num_width, 4)
 
     @property
     def max_instruction_width(self) -> int:
@@ -40,6 +48,10 @@ class PrettyPrinterBase:
     @property
     def max_byte_count(self) -> int:
         return self._max_byte_count
+
+    @property
+    def max_comment_width(self) -> int:
+        return self._max_comment_width
 
     def pretty_print(self) -> str:
         raise NotImplementedError